+++ RELEASE-NOTES.md
@@ -35,6 +35,8 @@ The project release numbers follow [Semantic Versioning](http://semver.org/spec/
   passed to ``target_link_options``.
 - For HIP-dependent builds, only add HCC include directory if it exists.
 - HIP CMake utilities updated to AMD's latest version
+- Updated ``add_code_coverage_target`` to ``blt_add_code_coverage_target``, which now supports
+  user-specified source directories
 
 ### Fixed
 - ClangFormat checks now support multiple Python executable names